Add duplicate non-topic ID rewrite to PDF2 merging.
A better place for this rewrite should be earlier in the preprocessing
chain or in MergeTopicParser; adding the rewrite requires
refactoring in the Java code to make it easier and unit tests to
guard against regression bugs. When the rewrite has been implemented in
preprocessing, changes in this commit should be removed.