Remove whitespace from menucascade - ID: 3423537
The preferred way would be to use xsl:strip-space to remove whitespace
from menucascade during tree build time, but it will not support
specialization. The next way is to not select menucascade for output,
option taken by the PDF processing. XHTML uses flag processing and makes
not choosing the text nodes difficult, thus it takes the approach of
having a template for menucascade text nodes and ignores them that way.