Fix unit tests on Windows
When running unit tests in Eclipse on Windows, unit test
is unable to delete temporary files. This is caused by running
Transformer by passing URIs as arguments. When the test code
runs the Transformer with InputStream and OutputStrean and the test
code closes the streams manually, temporary files can be deleted.