Add missing @class attribute to <navref> in chunking phase
A <topicset> element inherits a @chunk attribute value of "to-navigation" from
the DTD. In the chunking phase, elements with @chunk="to-navigation" are
converted into a <navref> element. Currently, this <navref> element gets no
@class attribute value.

With the Eclipse Help transtype, this <navref> element should be converted into
a <link> element in the Eclipse Help TOC file, but since map2eclipseImpl.xsl
(correctly) matches against the @class attribute instead of the element name, it
ignores the <navref> element and the <link> element is never created. Adding the
missing @class attribute to the <navref> element solves the problem.