Fix topicmerge topic duplication error #1539
In the startElement() method, the code checked whether a topicref has
already been visited by checking whether visitSet contains the (possibly
URL-encoded) @href of that topicref. However, it would add the URL-
decoded value of the @href into visitSet to mark a topicref as visited.
For example, it would add "lawn%20mower.xml" into visitSet but would
later check it for "lawn mower.xml" instead.

The upshot is that if the parser encountered a reference to the same
topic (for example inside a reltable), it would consider that topic
unvisited and include it in the output again.