Fix inconsistent translations
Noticed in a discussion with @simalot and @keberlein that Chapter and Part are defined differently for the TOC than they are in the content. The strings used in the content are correct, and we should use the same value in each spot. The correct usage has text before and after the number.