Use Sass `// comment` syntax per #2218
This PR replaces the C-style (multi-line) `/* comment */` syntax used to mark inline/block elements in the generated Sass partials for DITA domains, as these comments are passed through to the output even when empty rules that follow them are dropped by the compiler.

Switching these to Sass C++-style (single-line) `// comment` syntax reduces noise in the CSS output.

The old CSS comments have been preserved in the Sass source and are passed through to the CSS output on the assumption that they may be useful to users developing custom stylesheets.