Remove note about future licenses
As pointed out by SyncroSoft, language added to the license.txt file when we had dual licenses is potentially confusing and seems out of sync with the straightforward declaration (in many other locations) that we use the Apache 2.0 license. Removing the extra sentence about alternate licenses, to keep this instance as simple as others.

This should help clarify that if future releases use additional or alternate licenses, those licenses will apply only to those future versions.