Use arabic numbering for notices in back matter
We noticed that if a book map has `<notices>` in the back matter, the page numbering switches to roman numerals (under the assumption that this is the Notices section in the front matter). This is normally the correct default for any topic under `<notices>`, which is most commonly located in the front matter, and thus the page number format is picked up from the attribute set `page-sequence.notice`.

By default the body and back matter use arabic numerals; this fix checks to see if the notices is inside `<backmatter>`, and if so, switches format back to arabic.

Feel free to suggest a better way to do this; the current fix uses the same key from later in the template to pick up `<notices>` and check ancestry, which seems straightforward but looks cumbersome somehow.