Use Sass `// comment` syntax for partials copyrights
While testing the Sass changes in #2499, I noticed an unintended side-effect of the copyright statement revisions from #2467, which results in  multiple redundant copyright comments in the generated CSS files for HTML5.

This PR replaces the C-style (multi-line) `/* comment */` syntax used in the copyright statements of the referenced Sass partials, as these comments are all passed through to the output.

Switching these to Sass C++-style (single-line) `// comment` syntax reduces noise in the CSS output, like in #2369.

The copyright statements in the master Sass files `commonltr.scss` & `commonrtl.scss` are left in C-style syntax to ensure they are propagated to the generated CSS.

The resulting generated CSS files are functionally equivalent to those shipped with DITA-OT 2.3.3.

Signed-off-by: Roger Sheen <roger@infotexture.net>