Retain DITA source format
Module tried to retain the format of non-XML DITA topics by pushing them
into Job object, but temporary filename generator had not been created yet.
Adding a map that keeps track of the formats. A better solution would be retain
the format information with the source URIs as Reference objects, but that would
require larger refactoring.