Restore PR template
When issue templates were refactored for #3061 to support [multiple issue templates](https://blog.github.com/2018-01-25-multiple-issue-and-pull-request-templates/), the PR template was moved to the wrong location in https://github.com/dita-ot/dita-ot/commit/32df7f9.

This commit moves the template to the correct location in the `.github/PULL_REQUEST_TEMPLATE` subfolder to ensure the template is once again used for new pull requests.

Signed-off-by: Roger Sheen <roger@infotexture.net>