Copyedit CLI help
When the `dita help` subcommand is run, the current output is inconsistent with regard to initial capitalization of each line. The majority of entries are parameter descriptions in sentence case, so this capitalizes the remaining arguments & options to align with this precedent, and clarifies a few descriptions in the process.

Signed-off-by: Roger Sheen <roger@infotexture.net>