Issue #3433 No warning for broken in-topic link
Added error DOTX032E. Seemed most relevant for this issue.
The error is raised only when the problematic xref doesn't have a text
node, and is not possible to reference in current topic. I don't know
if a warning should be created for this situation, I just found the
position where the error message could be generated for this issue.

Signed-off-by: Konstantinos Tsiligkiris <ktsiligkiris@outlook.com>