Remove obsolete `log4j` remnants
Up until DITA-OT 2.4, `log4j` was bundled as a dependency of the Apache™ Formatting Objects Processor (FOP).

The `log4j.jar` library was removed from `plugins/org.dita.pdf2.fop/fop/lib/` in DITA-OT 2.4, when FOP was upgraded to version 2.1 with #2399.

This commit removes the remaining `log4j` configuration files, which are not referenced elsewhere in the DITA-OT codebase.

Signed-off-by: Roger Sheen <roger@infotexture.net>