Simplify `docsrc/` exclude patterns in dist builds
Hidden files & folders were previously excluded from the distribution package by name, which allowed new dotfiles to creep in if not added explicitly here.

(This happened with `.git-blame-ignore-revs`, which shipped with 4.1.1 & 4.1.2 after it was moved to the repo root in dita-ot/docs@dd111ba.)

This replaces the existing path references with the same `.*` pattern that is used 5 lines above to exclude all dotfiles and hidden folders from the core repo.

Signed-off-by: Roger Sheen <roger@infotexture.net>