Fix trademark in related links in XHTML/eclipsehelp transtypes
Porting fix done on the HTML5 transform for dita-ot/dita-ot#2303 in PR https://github.com/dita-ot/dita-ot/pull/2394.

Message from original fix from @eerohele:
An intermediate tree is assigned to
the $unordered-links variable, which is then applied. The root node of an intermediate tree is not a document node, which causes any calls to the key() function to fail.

The solution is to pass the original root node as a tunnel parameter and
use that as the context node for any calls to key().

For issue dita-ot/dita-ot#4686

Signed-off-by: GDELORY <gdelory@sk.ibm.com>